pyproject.toml
[project]
name = "lametric-hass-local"
version = "0.1.0"
description = ""
authors = [
    { name = "Caspar Weber" }
]
dependencies = [
    "lametric-py>=1.0.0",
    "homeassistant>=2026.1.0",
]
requires-python = ">=3.14"
readme = "README.md"
license = { text = "MIT" }

# ------------------------------------------------------------------------------
# UV & DEV DEPENDENCIES
# ------------------------------------------------------------------------------
[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.9.0",
    "ruff>=0.3.0",
    "pre-commit>=3.6.0",
    "python-semantic-release>=9.0.0",
    "commitizen>=3.0.0",
]

# ------------------------------------------------------------------------------
# RUFF (Linter & Formatter)
# ------------------------------------------------------------------------------
[tool.ruff]
target-version = "py314"
line-length = 88

[tool.ruff.lint]
# E: Pycodestyle, F: Pyflakes, I: Isort (Imports), N: Naming, UP: Upgrade, B: Bugbear
select = ["E", "F", "I", "N", "UP", "B"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ------------------------------------------------------------------------------
# MYPY (Type Checking)
# ------------------------------------------------------------------------------
[tool.mypy]
python_version = "3.14"
strict = true
ignore_missing_imports = true
warn_unused_configs = true
disallow_untyped_defs = true

# ------------------------------------------------------------------------------
# PYTEST (Testing)
# ------------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "--cov=custom_components/lametric_hass_local --cov-report=term-missing"

# ------------------------------------------------------------------------------
# SEMANTIC RELEASE (Auto-Versioning)
# ------------------------------------------------------------------------------
[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
major_on_zero = false

[tool.semantic_release.changelog]
exclude_commit_patterns = ["chore", "ci"]

[tool.semantic_release.branches.main]
match = "main"
